separating mapping and optimization
